+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/UserRole.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = false)
-@Table(name = "user_role", uniqueConstraints = @UniqueConstraint(columnNames = {"role", "app_user_id", "inst_group_id"}))
+@Table(name = "user_role", uniqueConstraints = @UniqueConstraint(columnNames = {"role", "app_user_id", "institution_id"}))
 public class UserRole extends CreationAudited {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,10 +31,10 @@ public class UserRole extends CreationAudited {
     private AppUser user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "inst_group_id")
+    @JoinColumn(name = "institution_id")
     @NotNull
     @ToString.Exclude
     @EqualsAndHashCode.Include
-    private Group group;
+    private Institution institution;
 
 }