+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/InstitutionController.java
@@ -6,24 +6,33 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import lombok.val;
 import org.springdoc.core.converters.models.PageableAsQueryParam;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 import pl.cyfronet.s4e.bean.AppRole;
+import pl.cyfronet.s4e.controller.request.AddMemberRequest;
 import pl.cyfronet.s4e.controller.request.CreateChildInstitutionRequest;
 import pl.cyfronet.s4e.controller.request.CreateInstitutionRequest;
 import pl.cyfronet.s4e.controller.request.UpdateInstitutionRequest;
 import pl.cyfronet.s4e.controller.response.BasicInstitutionResponse;
 import pl.cyfronet.s4e.controller.response.InstitutionResponse;
+import pl.cyfronet.s4e.controller.response.MemberResponse;
+import pl.cyfronet.s4e.event.OnAddToInstitutionEvent;
+import pl.cyfronet.s4e.event.OnRemoveFromInstitutionEvent;
 import pl.cyfronet.s4e.ex.InstitutionCreationException;
 import pl.cyfronet.s4e.ex.NotFoundException;
 import pl.cyfronet.s4e.ex.S3ClientException;
 import pl.cyfronet.s4e.security.AppUserDetails;
+import pl.cyfronet.s4e.service.AppUserService;
 import pl.cyfronet.s4e.service.InstitutionService;
 import pl.cyfronet.s4e.util.AppUserDetailsSupplier;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Set;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static pl.cyfronet.s4e.Constants.API_PREFIX_V1;
@@ -34,6 +43,8 @@ import static pl.cyfronet.s4e.Constants.API_PREFIX_V1;
 @Tag(name = "institution", description = "The Institution API")
 public class InstitutionController {
     private final InstitutionService institutionService;
+    private final AppUserService appUserService;
+    private final ApplicationEventPublisher eventPublisher;
 
     @Operation(summary = "Create a new institution")
     @ApiResponses({
@@ -110,6 +121,55 @@ public class InstitutionController {
     public InstitutionResponse get(@PathVariable("institution") String institutionSlug) throws NotFoundException {
         return institutionService.findBySlug(institutionSlug, InstitutionResponse.class)
                 .orElseThrow(() -> new NotFoundException("Institution not found for id '" + institutionSlug + "'"));
+    }
+
+    @Operation(summary = "Get members")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved members"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @GetMapping("/institutions/{institution}/members")
+    public Set<MemberResponse> getMembers(@PathVariable("institution") String institutionSlug) {
+        return institutionService.getMembers(institutionSlug, MemberResponse.class);
+    }
+
+    @Operation(summary = "Add a new member to the group")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "If member was added"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @PostMapping(value = "/institutions/{institution}/members", consumes = APPLICATION_JSON_VALUE)
+    public void addMember(@RequestBody AddMemberRequest request,
+                          @PathVariable("institution") String institutionSlug)
+            throws NotFoundException {
+        institutionService.addMember(institutionSlug, request.getEmail());
+        eventPublisher.publishEvent(
+                new OnAddToInstitutionEvent(request.getEmail(), institutionSlug, LocaleContextHolder.getLocale()));
+    }
+
+    @Operation(summary = "Remove a member from the group")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "If member was removed"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @DeleteMapping("/institutions/{institution}/members/{id}")
+    public void removeMember(@PathVariable("institution") String institutionSlug,
+                             @PathVariable Long id)
+            throws NotFoundException {
+        val appUser = appUserService.findById(id)
+                .orElseThrow(() -> new NotFoundException("User not found for id: '" + id + "'"));
+        institutionService.removeMember(institutionSlug, appUser.getEmail());
+        eventPublisher.publishEvent(
+                new OnRemoveFromInstitutionEvent(appUser.getEmail(), institutionSlug, LocaleContextHolder.getLocale()));
     }
 
     @Operation(summary = "Delete an institution")