+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/request/AddMemberRequest.java
@@ -1,0 +1,13 @@
+package pl.cyfronet.s4e.controller.request;
+
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.constraints.Email;
+
+@Data
+@Builder
+public class AddMemberRequest {
+    @Email
+    private String email;
+}