+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/UserRoleResponse.java
@@ -7,9 +7,6 @@ public interface UserRoleResponse {
 
     AppRole getRole();
 
-    @Value("#{target.group.slug}")
-    String getGroupSlug();
-
-    @Value("#{target.group.institution.slug}")
+    @Value("#{target.institution.slug}")
     String getInstitutionSlug();
 }