+++ s4e-backend/src/test/java/pl/cyfronet/s4e/controller/InstitutionControllerTest.java
@@ -3,6 +3,7 @@ package pl.cyfronet.s4e.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -11,15 +12,15 @@ import org.springframework.test.web.servlet.MockMvc;
 import pl.cyfronet.s4e.BasicTest;
 import pl.cyfronet.s4e.TestDbHelper;
 import pl.cyfronet.s4e.bean.*;
+import pl.cyfronet.s4e.controller.request.AddMemberRequest;
 import pl.cyfronet.s4e.data.repository.AppUserRepository;
-import pl.cyfronet.s4e.data.repository.GroupRepository;
 import pl.cyfronet.s4e.data.repository.InstitutionRepository;
 import pl.cyfronet.s4e.data.repository.UserRoleRepository;
 import pl.cyfronet.s4e.service.SlugService;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static pl.cyfronet.s4e.Constants.API_PREFIX_V1;
@@ -41,9 +42,6 @@ public class InstitutionControllerTest {
     private UserRoleRepository userRoleRepository;
 
     @Autowired
-    private GroupRepository groupRepository;
-
-    @Autowired
     private ObjectMapper objectMapper;
 
     @Autowired
@@ -56,6 +54,7 @@ public class InstitutionControllerTest {
     private MockMvc mockMvc;
 
     private AppUser appUser;
+    private AppUser instAdmin;
     private AppUser admin;
 
     private String testInstitution = "Test Institution";
@@ -81,15 +80,24 @@ public class InstitutionControllerTest {
                 .admin(true)
                 .build());
 
+        instAdmin = appUserRepository.save(AppUser.builder()
+                .email("noop@mail.pl")
+                .name("Get")
+                .surname("Profile")
+                .password("{noop}password")
+                .enabled(true)
+                .build());
+
         slugInstitution = slugService.slugify(testInstitution);
         Institution institution = institutionRepository.save(Institution.builder()
                 .name(testInstitution)
                 .slug(slugInstitution)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
-        UserRole userRole = UserRole.builder().role(AppRole.GROUP_MEMBER).user(appUser).group(group).build();
+        UserRole userRole = UserRole.builder().role(AppRole.INST_ADMIN).user(instAdmin).institution(institution).build();
+        UserRole userRole2 = UserRole.builder().role(AppRole.GROUP_MEMBER).user(instAdmin).institution(institution).build();
         userRoleRepository.save(userRole);
+        userRoleRepository.save(userRole2);
     }
 
     @AfterEach
@@ -101,7 +109,7 @@ public class InstitutionControllerTest {
     public void shouldGetUserInstitution() throws Exception {
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions")
                 .contentType(MediaType.APPLICATION_JSON)
-                .with(jwtBearerToken(appUser, objectMapper)))
+                .with(jwtBearerToken(instAdmin, objectMapper)))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$[0].name").value(is(equalTo(testInstitution))))
                 .andExpect(jsonPath("$[0].slug").value(is(equalTo(slugInstitution))));
@@ -132,9 +140,8 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
-        UserRole userRole = UserRole.builder().role(AppRole.GROUP_MEMBER).user(appUser).group(group).build();
+        UserRole userRole = UserRole.builder().role(AppRole.GROUP_MEMBER).user(appUser).institution(institution).build();
         userRoleRepository.save(userRole);
 
         String testInstitutionZK_MAZ = "Test Institution ZK - Mazowieckie";
@@ -144,9 +151,8 @@ public class InstitutionControllerTest {
                 .slug(slugInstitutionZK_MAZ)
                 .parent(institution)
                 .build());
-        Group group2 = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution2).build());
 
-        userRole = UserRole.builder().role(AppRole.GROUP_MEMBER).user(appUser).group(group2).build();
+        userRole = UserRole.builder().role(AppRole.GROUP_MEMBER).user(appUser).institution(institution2).build();
         userRoleRepository.save(userRole);
 
         String testInstitutionPAK = "Test Institution PAK";
@@ -155,13 +161,12 @@ public class InstitutionControllerTest {
                 .name(testInstitutionPAK)
                 .slug(slugInstitutionPAK)
                 .build());
-        groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions")
                 .contentType(MediaType.APPLICATION_JSON)
                 .with(jwtBearerToken(appUser, objectMapper)))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.length()", is(equalTo(3))));
+                .andExpect(jsonPath("$.length()", is(equalTo(2))));
     }
 
     @Test
@@ -172,7 +177,6 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         String testInstitutionZK_MAZ = "Test Institution ZK - Mazowieckie";
         String slugInstitutionZK_MAZ = slugService.slugify(testInstitutionZK_MAZ);
@@ -181,7 +185,6 @@ public class InstitutionControllerTest {
                 .slug(slugInstitutionZK_MAZ)
                 .parent(institution)
                 .build());
-        Group group2 = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution2).build());
 
         String testInstitutionPAK = "Test Institution PAK";
         String slugInstitutionPAK = slugService.slugify(testInstitutionPAK);
@@ -189,7 +192,6 @@ public class InstitutionControllerTest {
                 .name(testInstitutionPAK)
                 .slug(slugInstitutionPAK)
                 .build());
-        groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions")
                 .contentType(MediaType.APPLICATION_JSON)
@@ -206,7 +208,6 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}", slugInstitutionZK)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -224,7 +225,6 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}", slugInstitutionZK)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -242,7 +242,6 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{intitution}", slugInstitutionZK)
                 .contentType(MediaType.APPLICATION_JSON))
@@ -264,10 +263,124 @@ public class InstitutionControllerTest {
                 .name(testInstitutionZK)
                 .slug(slugInstitutionZK)
                 .build());
-        Group group = groupRepository.save(Group.builder().name("__default__").slug("default").institution(institution).build());
 
         mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{intitution}", slugInstitutionZK)
                 .with(jwtBearerToken(appUser, objectMapper)))
                 .andExpect(status().isForbidden());
+    }
+
+    @Nested
+    class GetMembers {
+        @Test
+        public void shouldWork() throws Exception {
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper)))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.length()", is(equalTo(1))));
+        }
+
+        @Test
+        public void shouldReturnEmptyIfNoMembers() throws Exception {
+            String testInstitutionZK = "Test Institution ZK";
+            String slugInstitutionZK = slugService.slugify(testInstitutionZK);
+            Institution institution = institutionRepository.save(Institution.builder()
+                    .name(testInstitutionZK)
+                    .slug(slugInstitutionZK)
+                    .build());
+
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitutionZK)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(admin, objectMapper)))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.length()", is(equalTo(0))));
+        }
+
+        @Test
+        public void shouldBeSecured() throws Exception {
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(appUser, objectMapper)))
+                    .andExpect(status().isForbidden());
+        }
+    }
+
+    @Nested
+    class AddMembers {
+        @Test
+        public void shouldWork() throws Exception {
+            AddMemberRequest request = AddMemberRequest.builder().email(appUser.getEmail()).build();
+            mockMvc.perform(post(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper))
+                    .content(objectMapper.writeValueAsBytes(request)))
+                    .andExpect(status().isOk());
+
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper)))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.length()", is(equalTo(2))));
+        }
+
+        @Test
+        public void shouldBeSecured() throws Exception {
+            AddMemberRequest request = AddMemberRequest.builder().build();
+            mockMvc.perform(post(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(appUser, objectMapper))
+                    .content(objectMapper.writeValueAsBytes(request)))
+                    .andExpect(status().isForbidden());
+        }
+    }
+
+    @Nested
+    class RemoveMembers {
+        @Test
+        public void shouldWork() throws Exception {
+            AddMemberRequest request = AddMemberRequest.builder().email(appUser.getEmail()).build();
+            mockMvc.perform(post(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper))
+                    .content(objectMapper.writeValueAsBytes(request)))
+                    .andExpect(status().isOk());
+
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper)))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.length()", is(equalTo(2))));
+
+            mockMvc.perform(delete(API_PREFIX_V1 + "/institutions/{institution}/members/{id}", slugInstitution, appUser.getId())
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper)))
+                    .andExpect(status().isOk());
+
+            mockMvc.perform(get(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper)))
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath("$.length()", is(equalTo(1))));
+        }
+
+        @Test
+        public void shouldBeSecured() throws Exception {
+            mockMvc.perform(delete(API_PREFIX_V1 + "/institutions/{institution}/members/{id}", slugInstitution, appUser.getId())
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(appUser, objectMapper)))
+                    .andExpect(status().isForbidden());
+
+            AddMemberRequest request = AddMemberRequest.builder().email(appUser.getEmail()).build();
+            mockMvc.perform(post(API_PREFIX_V1 + "/institutions/{institution}/members", slugInstitution)
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(instAdmin, objectMapper))
+                    .content(objectMapper.writeValueAsBytes(request)))
+                    .andExpect(status().isOk());
+
+            mockMvc.perform(delete(API_PREFIX_V1 + "/institutions/{institution}/members/{id}", slugInstitution, appUser.getId())
+                    .contentType(MediaType.APPLICATION_JSON)
+                    .with(jwtBearerToken(appUser, objectMapper)))
+                    .andExpect(status().isForbidden());
+        }
     }
 }